Splitted the reader test to own file.
